refactor: remove comment, rename file
